Move skill tool workers onto the shared loop
Replaces skillworker's own loop, TaskRunner and conductorRunner with an adapter
onto taskworker, leaving this package owning just the tool logic and the
{"result": …} envelope.

ToolHandler is kept rather than collapsed into taskworker.Handler. Seven
constructors return it, three functions pass it as map[string]ToolHandler, and
ten tests assert against it; adapting instead is one line at startSkillWorkers
and avoids duplicating the result wrap across every tool. It is a cohesive local
shape — tool logic in, raw JSON out — and stays this package's vocabulary.

Preserved: the {skillName}__{tool} task type, the conductor-cli worker id, one
task per poll, and wrapResult's fallbacks — non-JSON output carried through as a
string, and nil output yielding {"result": null}.

The loop and runner tests are deleted rather than kept: they covered the code
being removed, and taskworker's tests cover its replacement. handlers_test.go is
untouched and remains the regression guard for the tool logic.
